JumpScale9Lib/servers/grid_capacity/server/requirements.txt
# DO NOT EDIT THIS FILE. This file will be overwritten when re-running go-raml.


Flask==0.12.2
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
itsdangerous==0.24
jsonschema==2.6.0
six==1.11.0
python-jose==2.0.2
pycrypto==2.6.1
pycryptodome==3.5.1
click==6.7
ecdsa==0.13
future==0.16.0
flask-mongoengine==0.9.5
pycountry==18.5.26
pycountry-convert==0.7.2